# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankalpa Odisha API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [master-data]
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createDepartment
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NameRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /departments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateDepartment
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DepartmentResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteDepartment
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /categories:
    get:
      operationId: listCategories
      tags: [master-data]
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createCategory
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NameRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /categories/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateCategory
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteCategory
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /tags:
    get:
      operationId: listTags
      tags: [master-data]
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createTag
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NameRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /tags/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateTag
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteTag
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /districts:
    get:
      operationId: listDistricts
      tags: [master-data]
      responses:
        "200":
          description: Districts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createDistrict
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NameRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /districts/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateDistrict
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DistrictResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteDistrict
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /blocks:
    get:
      operationId: listBlocks
      tags: [master-data]
      parameters:
        - in: query
          name: districtId
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Blocks
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createBlock
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlockCreateRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockResponse" }
        "400": { $ref: "#/components/responses/ErrorResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /blocks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateBlock
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BlockResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteBlock
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
  /occasions:
    get:
      operationId: listOccasions
      tags: [master-data]
      responses:
        "200":
          description: Occasions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccasionListResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createOccasion
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NameRequest" }
      responses:
        "200":
          description: Existing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccasionResponse" }
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccasionResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
  /occasions/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateOccasion
      tags: [master-data]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MasterUpdateRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OccasionResponse" }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
        "409": { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteOccasion
      tags: [master-data]
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/ErrorResponse" }
        "403": { $ref: "#/components/responses/ErrorResponse" }
        "404": { $ref: "#/components/responses/ErrorResponse" }
...
components:
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    NameRequest:
      type: object
      properties:
        name: { type: string }
      required: [name]
    MasterUpdateRequest:
      description: At least one of `name` or `isActive` must be provided.
      type: object
      minProperties: 1
      properties:
        name: { type: string }
        isActive: { type: boolean }
    BlockCreateRequest:
      type: object
      properties:
        name: { type: string }
        districtId: { type: integer }
      required: [name, districtId]
    Department:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
      required: [id, name, isActive]
    District:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
      required: [id, name, isActive]
    DistrictListResponse:
      type: object
      properties:
        districts:
          type: array
          items: { $ref: "#/components/schemas/District" }
      required: [districts]
    DistrictResponse:
      type: object
      properties:
        district: { $ref: "#/components/schemas/District" }
      required: [district]
...
    BlockListResponse:
      type: object
      properties:
        blocks:
          type: array
          items: { $ref: "#/components/schemas/Block" }
      required: [blocks]
    BlockResponse:
      type: object
      properties:
        block: { $ref: "#/components/schemas/Block" }
      required: [block]
...
    OccasionListResponse:
      type: object
      properties:
        occasions:
          type: array
          items: { $ref: "#/components/schemas/Occasion" }
      required: [occasions]
    OccasionResponse:
      type: object
      properties:
        occasion: { $ref: "#/components/schemas/Occasion" }
      required: [occasion]
...
    DepartmentListResponse:
      type: object
      properties:
        departments:
          type: array
          items: { $ref: "#/components/schemas/Department" }
      required: [departments]
    DepartmentResponse:
      type: object
      properties:
        department: { $ref: "#/components/schemas/Department" }
      required: [department]
    Category:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
      required: [id, name, isActive]
    CategoryListResponse:
      type: object
      properties:
        categories:
          type: array
          items: { $ref: "#/components/schemas/Category" }
      required: [categories]
    CategoryResponse:
      type: object
      properties:
        category: { $ref: "#/components/schemas/Category" }
      required: [category]
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        isActive: { type: boolean }
      required: [id, name, isActive]
    TagListResponse:
      type: object
      properties:
        tags:
          type: array
          items: { $ref: "#/components/schemas/Tag" }
      required: [tags]
    TagResponse:
      type: object
      properties:
        tag: { $ref: "#/components/schemas/Tag" }
      required: [tag]
...